app/requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.6
blinker==1.4
cffi==1.12.2
Click==7.0
cryptography==2.6.1
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
flask-marshmallow==0.10.0
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
marshmallow==2.19.1
Pillow==5.4.1
pycparser==2.19
PyJWT==1.7.1
six==1.12.0
SQLAlchemy==1.3.1
Werkzeug==0.15.1
